Parse options for -XX:jaasServer: & refactor

+++ runtime/compiler/env/J9PersistentInfo.hpp
@@ -290,8 +290,8 @@ class PersistentInfo : public OMR::PersistentInfoConnector
    void setJaasMode(JaasModes m) { _jaasMode = m; }
    std::string getJaasServerAddress() const { return _jaasServerAddress; }
    void setJaasServerAddress(char *addr) { _jaasServerAddress = addr; }
-   uintptr_t getJaasServerPort() const { return _jaasServerPort; }
-   void setJaasServerPort(uintptr_t port) { _jaasServerPort = port; }
+   uint32_t getJaasServerPort() const { return _jaasServerPort; }
+   void setJaasServerPort(uint32_t port) { _jaasServerPort = port; }
    std::string getJaasServerConnectionInfo() const { return _jaasServerAddress + ":" + std::to_string(_jaasServerPort); }
 
    private:
@@ -378,7 +378,7 @@ class PersistentInfo : public OMR::PersistentInfoConnector
    int32_t _numLoadedClasses; ///< always increasing
    JaasModes _jaasMode; // NONJAAS_MODE, CLIENT_MODE, SERVER_MODE
    std::string _jaasServerAddress;
-   uintptr_t _jaasServerPort;
+   uint32_t _jaasServerPort;
    };
 
 }
